Combine multiple individual grid simulation runs into one

Combine multiple individual grid simulation runs into one

Usage

# S3 method for character
combine_data(...)

Arguments

...

Either a list of objects of the class demografr_abc_sims as produced by the function simulate_abc, or individual objects of this class given as standard function arguments, or paths to 'rds' files containing serializations of such demografr_abc_sims objects.

Value

A combined object of the class demografr_abc_sims