Extract inferred posteriors in a tabular format

Extract inferred posteriors in a tabular format

Usage

extract_posterior(abc, param = NULL, posterior = c("adj", "unadj"))

Arguments

abc

An object produced by run_abc

param

A character vector containing either parameter names to summarize, or a regex-like matches to be used for subsetting. If NULL (the default), all parameters will be extracted.

posterior

Should an "adj"-usted or "unadj"-usted posterior be extracted? (Default is "adj").

Value

A data frame in the long format with posterior values of parameters

Examples

# read example ABC result with an inferred joint posterior distribution
abc_res <- readRDS(system.file("examples/basics_abc.rds", package = "demografr"))

# extract the entire posterior sample for all parameters as a data frame
extract_posterior(abc_res)
#> # A tibble: 800 × 2
#>    param     value
#>    <chr>     <dbl>
#>  1 Ne_A  1714.    
#>  2 Ne_B  1047.    
#>  3 Ne_C  9053.    
#>  4 Ne_D  3855.    
#>  5 T_AB  2048.    
#>  6 T_BC  6082.    
#>  7 T_CD  8075.    
#>  8 gf_BC    0.0247
#>  9 Ne_A  2130.    
#> 10 Ne_B   787.    
#> # ℹ 790 more rows

# extract the posterior sample for one parameter
extract_posterior(abc_res, param = "Ne_A")
#> # A tibble: 100 × 2
#>    param value
#>    <chr> <dbl>
#>  1 Ne_A  1714.
#>  2 Ne_A  2130.
#>  3 Ne_A  2202.
#>  4 Ne_A  1898.
#>  5 Ne_A  1980.
#>  6 Ne_A  2230.
#>  7 Ne_A  1821.
#>  8 Ne_A  1989.
#>  9 Ne_A  1941.
#> 10 Ne_A  2032.
#> # ℹ 90 more rows

# extract posterior samples for parameters matching a regex
extract_posterior(abc_res, param = "^Ne_")
#> # A tibble: 400 × 2
#>    param value
#>    <chr> <dbl>
#>  1 Ne_A  1714.
#>  2 Ne_B  1047.
#>  3 Ne_C  9053.
#>  4 Ne_D  3855.
#>  5 Ne_A  2130.
#>  6 Ne_B   787.
#>  7 Ne_C  8553.
#>  8 Ne_D  3665.
#>  9 Ne_A  2202.
#> 10 Ne_B   851.
#> # ℹ 390 more rows